Start of ui-server
